Added basic SAM authorization
In Calhoun, the service this is based on, the /api/convert endpoint is authorized via a call to the SAM authorization service that powers Terra. I'm not entirely sure this is necessary- the user is only getting access to a reformated version of data they are sending in. Perhaps checking that the user is registered is a way of protecting from automated DOS attacks. Perhaps it was done to avoid changes to the security boundary for compliance. I will follow up here. For now, I am calling the same authorization service endpoint.

Added basic call to the SAM authorization service.
I don't love the elifs for handling the response out of SAM. Better error handling TBD

Also TBD is parameterizing the SAM root from config.

Updated readme